handle errors thrown by final "done" handler
The problem was that if you have a final handler that takes an error and just
throws it, the error ends up at the previous next() catch where it's just
bubbled up again to the "done" handler. This repeats for exactly the number of
handlers on the stack. The solution is to catch errors from the "done" handler
and mark them as _punt = true. This tells the next() catch statements to simply
rethrow the error (bubbling it out of the stack) instead of going back up the
next() stack and rethrowing again and again.